Crate xarray

Structs

ArcRef

A type that represents &'a Arc<T>.

BoxRef

A type that represents &'a Box<T>.

Cursor

A Cursor can traverse in the XArray by setting or increasing the target index and can perform read-only operations to the target item represented by the target index.

CursorMut

A CursorMut can traverse in the XArray by setting or increasing the target index and can perform read-write operations to the target item represented by the target index.

Range

An iterator over a range of entries in an XArray.

XArray

XArray is an abstract data type functioning like an expansive array of items where each item must be an 8-byte object, such as Arc<T> or Box<T>.

Enums

XMark

The mark type used in the XArray.

Traits

ItemEntry

A trait for the types users wish to store in an XArray.