xarray

Struct LockedXArray

pub struct LockedXArray<'a, P, M = NoneMark, G = PreemptDisabled>
where
    P: NonNullPtr + Send + Sync,
    G: SpinGuardian,
{ /* private fields */ }

The locked XArray which obtains its inner spinlock.

The locked XArray is able to create CursorMut and do mutable operations. There can only be one locked XArray at the same time.

Implementations

impl<P: NonNullPtr + Send + Sync, M, G: SpinGuardian> LockedXArray<'_, P, M, G>

pub fn clear(&mut self)

Clears the corresponding XArray.

pub fn cursor_mut(&mut self, index: u64) -> CursorMut<'_, P, M>

Creates a CursorMut to perform read- and write-related operations.

pub fn store(&mut self, index: u64, item: P)

Stores the provided item at the target index.

pub fn remove(&mut self, index: u64) -> Option<P::Ref<'_>>

Removes the item at the target index.

Returns the removed item if some item was previously stored in the same position.

pub fn cursor(&self, index: u64) -> Cursor<'_, P, M>

Creates a Cursor to perform read-related operations.

pub fn range(&self, range: Range<u64>) -> Range<'_, P, M>

Creates a Range to immutably iterated over the specified range.

pub fn load(&self, index: u64) -> Option<P::Ref<'_>>

Loads the index-th item.

If the target item exists, it will be returned with Some(_), otherwise, None will be returned.