Crate xarray

This crate introduces a RCU-based XArray implementation.

XArray is an abstract data type functioning like an expansive array of items where each item is a NonNullPtr, such as Arc<T> or Box<T>. It facilitates efficient sequential access to adjacent entries, supporting multiple concurrent reads and exclusively allowing one write operation at a time.

In addition to directly manipulating the XArray, users can typically achieve more flexible operations by creating a Cursor/CursorMut within the XArray. Since the XArray enforces a single write operation at any given time, performing write operations requires first acquiring a LockedXArray by calling its lock method.

XArray also provides a convenient way to mark individual items (see XMark).

Example

use alloc::sync::Arc;

use crare::rcu_xarray::*;
use crate::task::disable_preempt;

let xarray_arc: XArray<Arc<i32>> = XArray::new();
let value = Arc::new(10);
xarray_arc.lock().store(10, value);

let guard = disable_preempt();
assert_eq!(*xarray_arc.load(&guard, 10).unwrap().as_ref(), 10);

// Usage of the cursor

let locked_xarray = xarray_arc.lock();
let cursor_mut = locked_xarray.cursor_mut(100);

let value = Arc::new(100);
cursor_mut.store(value);
assert_eq!(cursor_mut.load(10).unwrap().as_ref(), 100);
let cursor = xarray_arc.cursor(&guard, 100);
assert_eq!(cursor.load(10).unwrap().as_ref(), 100);

Background

The XArray concept was originally introduced by Linux, which keeps the data structure of Linux Radix Trees.

Structs

Cursor

A Cursor can traverse in the XArray by setting or increasing the target index and can perform read-only operations to the target item.

CursorMut

A CursorMut can traverse in the XArray by setting or increasing the target index and can perform read-write operations to the target item.

LockedXArray

The locked XArray which obtains its inner spinlock.

Range

An iterator over a range of entries in an XArray.

SetMarkError

An error indicating that the mark cannot be set because the item does not exist.

XArray

A RCU-based XArray implementation.

Enums

XMark

The mark type used in the XArray.