xarray

Struct XArray

pub struct XArray<P, M = NoneMark>
where
    P: NonNullPtr + Send + Sync,
{ /* private fields */ }

A RCU-based XArray implementation.

XArray is used to store NonNullPtr, with the additional requirement that user-stored pointers must have a minimum alignment of 2 bytes.

XArray is RCU-based, which means:

Multiple concurrent readers are permitted.

Only one writer is allowed at a time.

Simultaneous read operations are allowed while writing.

Readers may see stale data (see Cursor and CursorMut for more information).

Interaction with XArray is generally through Cursor and CursorMut. Similar to XArray’s read-write properties, multiple Cursors may coexist (shared read access) and only one CursorMut may exist at a time (exclusive write access).

To create a Cursor, users can invoke XArray::cursor with an atomic-guard. To create a CursorMut, users need to call XArray::lock or XArray::lock_irq_disabled first to obtain a LockedXArray first.

XArray enables marking of individual items for user convenience. Items can have up to three distinct marks by default, with each mark independently maintained. Users can use self-defined types as marks by implementing the From<Type> trait for XMark. Marking is also applicable to internal nodes, indicating marked descendant nodes, though such marking is not transparent to users.

Implementations

impl<P: NonNullPtr + Send + Sync, M> XArray<P, M>

pub const fn new() -> Self

Makes a new, empty XArray.

pub fn lock(&self) -> LockedXArray<'_, P, M>

Acquires the lock to perform mutable operations.

pub fn lock_irq_disabled(&self) -> LockedXArray<'_, P, M, LocalIrqDisabled>

Acquires the lock with local IRQs disabled to perform mutable operations.

pub fn cursor<'a, G: AsAtomicModeGuard>( &'a self, guard: &'a G, index: u64, ) -> Cursor<'a, P, M>

Creates a Cursor to perform read-related operations.

pub fn range<'a, G: AsAtomicModeGuard>( &'a self, guard: &'a G, range: Range<u64>, ) -> Range<'a, P, M>

Creates a Range to immutably iterated over the specified range.

pub fn load<'a, G: AsAtomicModeGuard>( &'a self, guard: &'a G, index: u64, ) -> Option<P::Ref<'a>>

Loads the index-th item.

If the target item exists, it will be returned with Some(_), otherwise, None will be returned.